add gender and email_notifications

+++ users/models.py
@@ -5,17 +5,31 @@ import os
 class Profile(models.Model):
     user = models.OneToOneField(User,verbose_name='Користувач', on_delete=models.CASCADE)
     img = models.ImageField('Фото користувача', default='default.png', upload_to='user_images')
+    gender = models.CharField(
+        verbose_name='Стать',
+        max_length=20,
+        choices=[
+            ('male', 'Чоловіча'),
+            ('female', 'Жіноча'),
+        ],
+        default='male'
+    )
+    email_notifications = models.BooleanField(
+        verbose_name='Отримання повідомлень на пошту',
+        default=False
+    )
 
     def __str__(self):
         return f'Профіль користувача {self.user.username}'
 
     def save(self, *args, **kwargs):
-        old = Profile.objects.get(pk=self.pk)
-        if old.img and old.img != self.img:
-            if old.img.name != 'default.png' and os.path.isfile(old.img.path):
-                os.remove(old.img.path)
+        if self.pk:
+            old = Profile.objects.get(pk=self.pk)
+            if old.img and old.img != self.img:
+                if old.img.name != 'default.png' and os.path.isfile(old.img.path):
+                    os.remove(old.img.path)
 
-        super().save()
+        super().save(*args, **kwargs)
         image = Image.open(self.img.path)
         if image.height > 256 or image.width > 256:
             resize = (256, 256)
